+++ pyls/_utils.py
@@ -29,37 +29,6 @@ def camel_to_underscore(string):
     return ALL_CAP_RE.sub(r'\1_\2', s1).lower()
 
 
-def list_to_string(value):
-    return ",".join(value) if type(value) == list else value
-
-
-def race_hooks(hook_caller, pool, **kwargs):
-    """Given a pluggy hook spec, execute impls in parallel returning the first non-None result.
-
-    Note this does not support a lot of pluggy functionality, e.g. hook wrappers.
-    """
-    impls = hook_caller._nonwrappers + hook_caller._wrappers
-    log.debug("Racing hook impls for hook %s: %s", hook_caller, impls)
-
-    if not impls:
-        return None
-
-    def _apply(impl):
-        try:
-            return impl, impl.function(**kwargs)
-        except Exception:
-            log.exception("Failed to run hook %s", impl.plugin_name)
-            raise
-
-    # imap unordered gives us an iterator over the items in the order they finish.
-    # We have to be careful to set chunksize to 1 to ensure hooks each get their own thread.
-    # Unfortunately, there's no way to interrupt these threads, so we just have to leave them be.
-    for impl, result in pool.imap_unordered(_apply, impls, chunksize=1):
-        if result is not None:
-            log.debug("Hook from plugin %s returned: %s", impl.plugin_name, result)
-            return result
-
-
 def find_parents(root, path, names):
     """Find files matching the given names relative to the given path.
 
@@ -93,3 +62,55 @@ def find_parents(root, path, names):
 
     # Otherwise nothing
     return []
+
+
+def list_to_string(value):
+    return ",".join(value) if type(value) == list else value
+
+
+def merge_dicts(dict_a, dict_b):
+    """Recursively merge dictionary b into dictionary a.
+
+    If override_nones is True, then
+    """
+    def _merge_dicts_(a, b):
+        for key in set(a.keys()).union(b.keys()):
+            if key in a and key in b:
+                if isinstance(a[key], dict) and isinstance(b[key], dict):
+                    yield (key, dict(_merge_dicts_(a[key], b[key])))
+                elif b[key] is not None:
+                    yield (key, b[key])
+                else:
+                    yield (key, a[key])
+            elif key in a:
+                yield (key, a[key])
+            elif b[key] is not None:
+                yield (key, b[key])
+    return dict(_merge_dicts_(dict_a, dict_b))
+
+
+def race_hooks(hook_caller, pool, **kwargs):
+    """Given a pluggy hook spec, execute impls in parallel returning the first non-None result.
+
+    Note this does not support a lot of pluggy functionality, e.g. hook wrappers.
+    """
+    impls = hook_caller._nonwrappers + hook_caller._wrappers
+    log.debug("Racing hook impls for hook %s: %s", hook_caller, impls)
+
+    if not impls:
+        return None
+
+    def _apply(impl):
+        try:
+            return impl, impl.function(**kwargs)
+        except Exception:
+            log.exception("Failed to run hook %s", impl.plugin_name)
+            raise
+
+    # imap unordered gives us an iterator over the items in the order they finish.
+    # We have to be careful to set chunksize to 1 to ensure hooks each get their own thread.
+    # Unfortunately, there's no way to interrupt these threads, so we just have to leave them be.
+    for impl, result in pool.imap_unordered(_apply, impls, chunksize=1):
+        if result is not None:
+            log.debug("Hook from plugin %s returned: %s", impl.plugin_name, result)
+            return result
