+++ pyls/config/__init__.py
@@ -1,0 +1,1 @@
+# Copyright 2017 Palantir Technologies, Inc.
