+++ pyls/config/flake8_conf.py
@@ -8,7 +8,6 @@ from pyls._utils import find_parents
 log = logging.getLogger(__name__)
 
 CONFIG_KEY = 'flake8'
-USER_CONFIGS = ['~\.flake8'] if os.name == 'nt' else ['~/.config/flake8']
 PROJECT_CONFIGS = ['.flake8', 'setup.cfg', 'tox.ini']
 
 OPTIONS = [
@@ -31,11 +30,7 @@ class Flake8Config(ConfigSource):
         else:
             return os.path.join(self.xdg_home, 'flake8')
 
-    def project_config(self, path=None):
-        if path:
-            files = find_parents(self.root_path, path, PROJECT_CONFIGS)
-        else:
-            files = [os.path.join(self.root_path, filename) for filename in PROJECT_CONFIGS]
-
+    def project_config(self, document_path):
+        files = find_parents(self.root_path, document_path, PROJECT_CONFIGS)
         config = self.read_config_from_files(files)
         return _utils.parse_config(config, CONFIG_KEY, OPTIONS)
